Split packages per screen
