+++ composeApp/src/commonMain/kotlin/com/example/mokumokusolo/App.kt
@@ -1,7 +1,7 @@
 package com.example.mokumokusolo
 
 import androidx.compose.runtime.Composable
-import com.example.mokumokusolo.ui.screen.HomeScreen
+import com.example.mokumokusolo.ui.screen.home.HomeScreen
 import com.example.mokumokusolo.ui.theme.MokuMokuSoloTheme
 import org.jetbrains.compose.ui.tooling.preview.Preview
