+++ composeApp/src/commonMain/kotlin/com/example/mokumokusolo/ui/screen/home/ExpenditureList.kt
@@ -1,6 +1,5 @@
-package com.example.mokumokusolo.ui.screen
+package com.example.mokumokusolo.ui.screen.home
 
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -20,57 +19,56 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.example.mokumokusolo.model.App
+import com.example.mokumokusolo.model.Expenditure
 import com.example.mokumokusolo.ui.theme.MokuMokuSoloTheme
-import mokumokusolo.composeapp.generated.resources.Res
-import mokumokusolo.composeapp.generated.resources.duolingo
-import org.jetbrains.compose.resources.painterResource
 import org.jetbrains.compose.ui.tooling.preview.Preview
 
 @Composable
-fun AppList(appList: List<App>, modifier: Modifier = Modifier) {
+fun ExpenditureList(
+    expenditureList: List<Expenditure>,
+    modifier: Modifier = Modifier
+) {
     LazyColumn(
         modifier = modifier
     ) {
-        items(appList) { app ->
-            AppListItem(app)
+        items(expenditureList) { expenditure ->
+            ExpenditureListItem(expenditure)
         }
     }
 }
 
 @Preview
 @Composable
-private fun AppListPreview() {
+fun ExpenditureListPreview() {
     MokuMokuSoloTheme {
-        val sampleApps = listOf(
-            App(
+        val sampleExpenditures = listOf(
+            Expenditure(
                 id = 1,
-                name = "Duolingo",
-                image = painterResource(Res.drawable.duolingo),
-                revenue = 1000.0
+                name = "Netflix",
+                amount = 780.0
             ),
-            App(
+            Expenditure(
                 id = 2,
-                name = "Duolingo2",
-                image = painterResource(Res.drawable.duolingo),
-                revenue = 1500.0
+                name = "Amazon Prime",
+                amount = 600.0
             ),
-            App(
+            Expenditure(
                 id = 3,
-                name = "Duolingo3",
-                image = painterResource(Res.drawable.duolingo),
-                revenue = 2000.0
+                name = "Spotify",
+                amount = 1080.0
             )
         )
-        AppList(
-            appList = sampleApps
+        ExpenditureList(
+            expenditureList = sampleExpenditures
         )
     }
 }
 
-
 @Composable
-fun AppListItem(app: App, modifier: Modifier = Modifier) {
+fun ExpenditureListItem(
+    expenditure: Expenditure,
+    modifier: Modifier = Modifier
+) {
     ElevatedCard(
         modifier = modifier
             .width(300.dp)
@@ -83,27 +81,20 @@ fun AppListItem(app: App, modifier: Modifier = Modifier) {
         Row(
             modifier = Modifier
                 .fillMaxWidth()
-                .padding(horizontal = 8.dp, vertical = 16.dp),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.spacedBy(8.dp)
+                .padding(vertical = 8.dp, horizontal = 16.dp),
+            horizontalArrangement = Arrangement.spacedBy(8.dp),
+            verticalAlignment = Alignment.CenterVertically
         ) {
-            Image(
-                painter = painterResource(Res.drawable.duolingo),
-                contentDescription = "",
-                modifier = Modifier
-                    .width(40.dp)
-            )
             Text(
-                text = app.name,
+                text = expenditure.name,
                 modifier = Modifier.weight(1f),
                 fontSize = 16.sp,
                 fontWeight = FontWeight.Normal,
                 style = MaterialTheme.typography.bodySmall
             )
             Text(
-                text = "¥${app.revenue}",
+                text = "¥${expenditure.amount}",
                 color = Color.Gray,
-                fontSize = 16.sp,
                 style = MaterialTheme.typography.bodySmall
             )
         }
@@ -112,14 +103,13 @@ fun AppListItem(app: App, modifier: Modifier = Modifier) {
 
 @Preview
 @Composable
-fun AppListItemPreview() {
+private fun ExpenditureListItemPreview() {
     MokuMokuSoloTheme {
-        AppListItem(
-            app = App(
+        ExpenditureListItem(
+            expenditure = Expenditure(
                 id = 1,
-                name = "Duolingo",
-                image = painterResource(Res.drawable.duolingo),
-                revenue = 1000.0
+                name = "Netflix",
+                amount = 1000.0
             )
         )
     }
